+++ utils/authorization.js
@@ -1,5 +1,3 @@
-// middleware
-
 const jwt = require('jsonwebtoken');
 const dotenv = require('dotenv');
 dotenv.config();
